requirements.txt
openai==1.51.0
python-dotenv==1.0.1
requests==2.32.3
httpx==0.27.2
httpcore==1.0.5
fastapi==0.115.0
pydantic==2.9.2
uvicorn==0.30.6
psycopg2-binary==2.9.9
